+++ packages/components/utils/src/index.ts
@@ -1,6 +1,5 @@
 export { Portal } from './Portal/Portal';
 export type { PortalProps } from './Portal/Portal';
-export { useArrowKeyNavigation } from './ArrowKeyNavigation/useArrowKeyNavigation';
 export { useKeyboard } from './useKeyboard/useKeyboard';
 export type { UseKeyboardProps } from './useKeyboard/useKeyboard';
 export { getStringMatch } from './getStringMatch/getStringMatch';
